Rename confusing module naming baxter_io_devices -> baxter_external_devices. Rename redundant module naming joint_recorder -> recorder.
