+++ scripts/joint_position_joystick.py
@@ -35,7 +35,7 @@ import argparse
 import rospy
 
 import baxter_interface
-import baxter_io_devices
+import baxter_external_devices
 
 
 def rotate(l):
@@ -171,11 +171,11 @@ def main():
 
     joystick = None
     if args.joystick == 'xbox':
-        joystick = baxter_io_devices.joystick.XboxController()
+        joystick = baxter_external_devices.joystick.XboxController()
     elif args.joystick == 'logitech':
-        joystick = baxter_io_devices.joystick.LogitechController()
+        joystick = baxter_external_devices.joystick.LogitechController()
     elif args.joystick == 'ps3':
-        joystick = baxter_io_devices.joystick.PS3Controller()
+        joystick = baxter_external_devices.joystick.PS3Controller()
     else:
         parser.error("Unsupported joystick type '%s'" % (args.joystick))
