+++ scripts/joint_position_keyboard.py
@@ -33,7 +33,7 @@ Baxter RSDK Joint Position Example: keyboard
 import rospy
 
 import baxter_interface
-import baxter_io_devices
+import baxter_external_devices
 
 
 def map_keyboard():
@@ -90,7 +90,7 @@ def map_keyboard():
     done = False
     print("Controlling joints. Press ? for help, Esc to quit.")
     while not done and not rospy.is_shutdown():
-        c = baxter_io_devices.getch()
+        c = baxter_external_devices.getch()
         if c:
             #catch Esc or ctrl-c
             if c in ['\x1b', '\x03']:
