+++ src/baxter_examples/__init__.py
@@ -25,4 +25,4 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from .joint_recorder import JointRecorder
+from .recorder import JointRecorder
